% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lock_factors.R
\name{lock_factors}
\alias{lock_factors}
\title{Set factor levels based on row order}
\usage{
lock_factors(df, cols = NULL, rev = FALSE)
}
\arguments{
\item{df}{A \code{data.frame} object.}

\item{cols}{A character or character vector of the column names to apply the
factor leveling to. Defaults to \code{NULL} in which case all character or factor
type variables will be re-leveled.}

\item{rev}{Reverse the factor level? Defaults to \code{FALSE}}
}
\value{
The function returns a data frame.
}
\description{
Description
}
\keyword{factors,}
\keyword{levels}
